Richard -Finished event Video download, still have to save tags 	-Added Clip class to manage the saving of plist and source control

+++ FBTraining/Live2BenchNative/UtilityClasses/DownloadItem.h
@@ -29,8 +29,8 @@ typedef NS_OPTIONS(NSInteger, DownloadType) {
 @interface DownloadItem : NSObject <NSURLConnectionDataDelegate,NSURLConnectionDelegate>
 {
 
-    NSUInteger      _expectedBytes;
-    NSUInteger      _receivedBytes;
+    double          _expectedBytes;
+    double          _receivedBytes;
     NSString        * url;
     NSURLConnection * theConnection;
     NSString        * path;
